dialects: (builtin) make `IntAttr` 'core'
