+++ xdsl/dialects/builtin.py
@@ -20,6 +20,7 @@ from immutabledict import immutabledict
 from typing_extensions import Self, TypeVar, deprecated, override
 
 from xdsl.dialect_interfaces import OpAsmDialectInterface
+from xdsl.dialects.core import IntAttr
 from xdsl.ir import (
     Attribute,
     AttributeCovT,
@@ -320,25 +321,6 @@ FlatSymbolRefAttrConstr = MessageConstraint(
 
 FlatSymbolRefAttr = Annotated[SymbolRefAttr, FlatSymbolRefAttrConstr]
 """SymbolRef constrained to have an empty `nested_references` property."""
-
-
-@irdl_attr_definition
-class IntAttr(Data[int]):
-    name = "builtin.int"
-
-    @classmethod
-    def parse_parameter(cls, parser: AttrParser) -> int:
-        with parser.in_angle_brackets():
-            data = parser.parse_integer()
-            return data
-
-    def print_parameter(self, printer: Printer) -> None:
-        with printer.in_angle_brackets():
-            printer.print_string(f"{self.data}")
-
-    def __bool__(self) -> bool:
-        """Returns True if value is non-zero."""
-        return bool(self.data)
 
 
 @dataclass(frozen=True)
